requirements.txt
allure-pytest==2.14.3
allure-python-commons==2.14.3
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
greenlet==3.2.3
idna==3.10
iniconfig==2.1.0
MouseInfo==0.1.3
packaging==25.0
playwright==1.53.0
pluggy==1.6.0
PyAutoGUI==0.9.54
pyee==13.0.0
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==1.0.9
pyobjc-core==11.1
pyobjc-framework-Cocoa==11.1
pyobjc-framework-Quartz==11.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytest==8.4.1
pytest-base-url==2.1.0
pytest-playwright==0.7.0
python-dotenv==1.1.1
python-slugify==8.0.4
pytweening==1.2.0
requests==2.32.4
rubicon-objc==0.5.1
text-unidecode==1.3
typing_extensions==4.14.0
urllib3==2.5.0
